Factor out session-error-handling.tsx.
